requirements.txt
flask==3.0.2
flask-cors==4.0.0
pytesseract==0.3.10
PyMuPDF==1.23.26
openai==1.12.0
Pillow==10.2.0
python-dotenv==1.0.1 